Make /hopkins faster by exclude small by default; make it the default
the initial loading is very very slow because of haveing to wake up
heroku CORS proxy *and* fetching hundreds of countries from the JH api.
The fetching is done in parallel but due to the high number we are being
throttled and the time is towards 10 seconds range.

I dropped the fetching of more than half of the countries.
Unfortunetally right now the list is hardcoded to the countries that
have less than 20000 cases at the moment. (Feb 2021)

You can always manually remove or set to 0 the url param  'exclude-contries=1'
to get all of them.